+++ wineguide/views/RatingsCVC.h
@@ -13,7 +13,7 @@
 @property (weak, nonatomic) IBOutlet UIImageView *glassImageView;
 @property (nonatomic) BOOL isRedWine;
 
--(void)setupImageViewForGlassNumber:(int)glassNumber andRating:(float)rating;
+-(void)setupImageViewForGlassNumber:(NSInteger)glassNumber andRating:(float)rating;
 -(void)resetCell;
 
 @end
